CU-1vj0cht - Write JUnit tests -> added Junit test for user management, Enum ROLES

+++ backend/shelf-server/src/main/java/com/htec/shelfserver/annotation/Roles.java
@@ -4,13 +4,13 @@ import lombok.Getter;
 
 @Getter
 public enum Roles {
-    SUPER_ADMIN(1),
-    MODERATOR(2),
-    USER(3);
+    SUPER_ADMIN(1L),
+    MODERATOR(2L),
+    USER(3L);
 
-    private final Integer value;
+    private final Long value;
 
-    Roles(Integer value) {
+    Roles(Long value) {
         this.value = value;
     }
 }
